+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/mock/action/http/HttpForwardActionHandler.java
@@ -28,10 +28,7 @@ import de.gematik.test.tiger.mockserver.httpclient.NettyHttpClient;
 import de.gematik.test.tiger.mockserver.model.HttpRequest;
 import de.gematik.test.tiger.mockserver.model.HttpResponse;
 import io.netty.channel.Channel;
-import java.net.InetSocketAddress;
 import java.util.concurrent.CompletableFuture;
-import java.util.function.UnaryOperator;
-import javax.annotation.Nullable;
 import lombok.extern.slf4j.Slf4j;
 
 /*
@@ -47,11 +44,7 @@ public class HttpForwardActionHandler {
     this.httpClient = httpClient;
   }
 
-  public HttpForwardActionResult sendRequest(
-      HttpRequest request,
-      Channel incomingChannel,
-      @Nullable InetSocketAddress remoteAddress,
-      UnaryOperator<HttpResponse> overrideHttpResponse) {
+  public HttpForwardActionResult sendRequest(HttpRequest request, Channel incomingChannel) {
     try {
       // TODO(jamesdbloom) support proxying via HTTP2, for now always force into HTTP1
       return new HttpForwardActionResult(
@@ -60,9 +53,9 @@ public class HttpForwardActionHandler {
               new HttpRequestInfo(
                   incomingChannel,
                   hopByHopHeaderFilter.onRequest(request).setProtocol(null),
-                  remoteAddress)),
-          overrideHttpResponse,
-          remoteAddress);
+                  null)),
+          null,
+          null);
     } catch (Exception e) {
       log.error("exception forwarding request {}", request, e);
     }